fix reporting scripts names
